11.7

+++ main.go
@@ -120,7 +120,7 @@ func analyzeStats(stats []float64) {
 	if totalNetwork > 0 {
 		freeNetwork := (totalNetwork * 8 / 1e7) - (usedNetwork * 8 / 1e7)
 		if usedNetwork/totalNetwork > networkUsageLimit {
-			fmt.Printf("Network bandwidth usage high: %d Mbit/s available\n", int64(freeNetwork))
+			fmt.Printf("Network bandwidth usage high: %f Mbit/s available\n", freeNetwork)
 		}
 	}
 }
